+++ CHANGELOG.md
@@ -1,8 +1,12 @@
-# @foxglove/fox version history
+# create-foxglove-extension version history
+
+## 0.5.0
+
+- Renamed package to `create-foxglove-extension`
 
 ## 0.1.1
 
-- Fix the "fox create" command in the README tutorial
+- Fix the "create" command in the README tutorial
 
 ## 0.1.0